Fix compilation warnings within test files (#118)

+++ test/run-fail/sdiv-udiv-inequal.c
@@ -2,11 +2,11 @@
 #include "caffeine.h"
 #include <stdint.h>
 
-uint32_t sdiv(uint32_t x, uint32_t y) {
+uint32_t sdiv(int32_t x, int32_t y) {
   caffeine_assume(x != INT32_MIN || y != -1);
   caffeine_assume(y != 0);
 
-  return (int32_t)x / (int32_t)y;
+  return x / y;
 }
 
 uint32_t udiv(uint32_t x, uint32_t y) {
